ДЗ 2.3. Введение в Maven и Spring Boot. Переделал с учетом комментариев и разбора ДЗ.

+++ src/main/java/me/aksaev/calculator/CalculatorService.java
@@ -2,7 +2,10 @@ package me.aksaev.calculator;
 
 public interface CalculatorService {
     public int plus(int num1, int num2);
+
     public int minus(int num1, int num2);
+
     public int multiply(int num1, int num2);
-    public int divide(int num1, int num2);
+
+    public double divide(int num1, int num2);
 }
