+++ src/main/java/me/aksaev/calculator/CalculatorServiceImpl.java
@@ -16,8 +16,8 @@ public class CalculatorServiceImpl implements CalculatorService {
         return num1 * num2;
     }
 
-    public int divide(int num1, int num2) {
-        return num1 / num2;
+    public double divide(int num1, int num2) {
+        return num1 / (double) num2;
     }
-    
+
 }
